FEATURE: Load existing projects during initialization
Existing projects are now loaded by the KApplication during initialization.
To do so a helper function was added that lists all filenames in a
directory that match the provided extension filter. The application
searches for all .json files that are located in KSettings.projectFolder
and tries to parse it as a project definition.